project-root4 - Copy/api-gateway/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
pyjwt==2.8.0
python-dotenv==1.0.0
pydantic==2.5.0
minio==7.2.0
